change rule of deleting area :if there has user or location under this area, then, can't delete it

+++ src/main/java/ins/map/web/PrpAreaInfoAction.java
@@ -2,9 +2,13 @@ package ins.map.web;
 
 import ins.framework.common.Page;
 import ins.framework.common.QueryRule;
+import ins.map.schema.model.LocationInfo;
 import ins.map.schema.model.PrpAreaInfo;
+import ins.map.service.facade.LocationInfoService;
 import ins.map.service.facade.PrpAreaInfoService;
 import ins.platform.common.web.SinoMapBaseAction;
+import ins.platform.schema.model.PrpDuser;
+import ins.platform.service.facade.UserService;
 
 import java.util.Date;
 import java.util.List;
@@ -16,7 +20,8 @@ public class PrpAreaInfoAction extends SinoMapBaseAction {
 	 */
 	private static final long serialVersionUID = 1L;
 	private PrpAreaInfoService prpAreaInfoService;
-	
+	private UserService userService;
+	private LocationInfoService locationInfoService;
 	private PrpAreaInfo prpAreaInfo;
 	private List<PrpAreaInfo> prpAreaInfoList;
 	private int totalCount;
@@ -65,6 +70,16 @@ public class PrpAreaInfoAction extends SinoMapBaseAction {
 	public String delete() {
 		try {
 			String comCode = getRequest().getParameter("comCode");
+			List<PrpDuser> userList = this.userService.findUserListByComCode(comCode);
+			if(userList.size() > 0) {//can't delete comcode if there exist user under this comcode
+				renderJSON(feilure(getText("action.deleteFeilure.hasUser")));
+				return null;
+			}
+			List<LocationInfo> locationList = this.locationInfoService.findLocationByComCode(comCode);
+			if(locationList.size() > 0) {//can't delete comcode is there exist location under this comcode
+				renderJSON(feilure(getText("action.deleteFeilure.hasLocation")));
+				return null;
+			}
 			this.prpAreaInfoService.delete(comCode);
 			renderJSON(success(getText("action.deleteSuccess"),"areaManage","","",""));
 		}catch(Exception e) {
@@ -140,6 +155,21 @@ public class PrpAreaInfoAction extends SinoMapBaseAction {
 	public void setComCode(String comCode) {
 		this.comCode = comCode;
 	}
-	
+
+	public UserService getUserService() {
+		return userService;
+	}
+
+	public void setUserService(UserService userService) {
+		this.userService = userService;
+	}
+
+	public LocationInfoService getLocationInfoService() {
+		return locationInfoService;
+	}
+
+	public void setLocationInfoService(LocationInfoService locationInfoService) {
+		this.locationInfoService = locationInfoService;
+	}
 	
 }
